- moved RECTANGLE_WIDTH/HEIGHT to csmacros.h; csdialogs*.* is not used any more   by the SDL version - removed global resource file "Resources" - action flags are now of type uint32, not int32 or long - map checkpoints in terminals (circles) display properly - size of sound data was miscalculated for extended sound headers of stereo   sounds - fixed many places where it was assumed that sizeof(long) = sizeof(void *) = 4;   now it works on Alpha CPUs - building in a separate directory works
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@513 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Misc/Random.h
@@ -14,19 +14,19 @@
 
 struct GM_Random
 {
-	unsigned long z, w, jsr, jcong, t[256], x, y;
+	uint32 z, w, jsr, jcong, t[256], x, y;
 	unsigned char c;
 	
-	unsigned long znew() {return (z=36969*(z&65535)+(z>>16));}
-	unsigned long wnew() {return (w=18000*(w&65535)+(w>>16));}
-	unsigned long MWC() {return ((znew()<<16)+wnew());}
-	unsigned long SHR3() {return (jsr=(jsr=(jsr=jsr^(jsr<<17))^(jsr>>13))^(jsr<<5));}
-	unsigned long CONG() {return (jcong=69069*jcong+1234567);}
-	unsigned long KISS() {return ((MWC()^CONG())+SHR3());}
-	unsigned long LFIB4() {t[c]=t[c]+t[c+58]+t[c+119]+t[c+178]; return t[++c];}
-	unsigned long SWB() {t[c+237]=(x=t[c+15])-(y=t[c]+(x<y)); return t[++c];}
+	uint32 znew() {return (z=36969*(z&65535)+(z>>16));}
+	uint32 wnew() {return (w=18000*(w&65535)+(w>>16));}
+	uint32 MWC() {return ((znew()<<16)+wnew());}
+	uint32 SHR3() {return (jsr=(jsr=(jsr=jsr^(jsr<<17))^(jsr>>13))^(jsr<<5));}
+	uint32 CONG() {return (jcong=69069*jcong+1234567);}
+	uint32 KISS() {return ((MWC()^CONG())+SHR3());}
+	uint32 LFIB4() {t[c]=t[c]+t[c+58]+t[c+119]+t[c+178]; return t[++c];}
+	uint32 SWB() {t[c+237]=(x=t[c+15])-(y=t[c]+(x<y)); return t[++c];}
 	float UNI() {return (KISS()*2.328306e-10);}
-	float VNI() {return (((long) KISS())*4.656613e-10);}
+	float VNI() {return (((int32) KISS())*4.656613e-10);}
 
 	/*
 		Random seeds must be used to reset z, w, jsr, jcong and
